SPARK-11314 move YarnSchedulerBackend class into yarn module, move service extension out of the subclasses and into this class.
